Add ability to specify symbols that should not get chpldoc generation and test
This adds the ability to specify that a module, variable, function, or class or
record type should not have documentation generated for it through chpldoc, even
if the file it lives in is run through chpldoc.  This is done through the pragma
"no doc", which is attached before the symbol in question.  For example:

pragma "no doc"
class Foo {...}

will cause no documentation for the class Foo.  This is true even of functions
that Foo specified within itself which do not have this pragma guarding them.

The test case I have added ensures this works for modules, configs, classes, and
functions.  Methods and records fall out trivially.  It also ensures that what
would normally be considered a chpldoc comment will not attach to later symbols
in the case where that symbol was skipped (though I don't expect this to be an
issue).