Manually increment the reference count in side of array and domain these().
This is a workaround for early deletion of array and domain copies that become fields in
iterators.  The these() function strips off the wrap-record which normally tracks
ownership (by manipulating the reference count contined in the array or domain
implementation).  Since the iterator structure creates a reference to the array
implementation (rather than just copying the wrap-record), it should increment the array
(or domain) reference count in its "constructor" and decrement it in its destructor.

This is currently difficult, because the iterator machinery is filled in after
resolution.  The change in the module code is a workaround until more of the iterator
machinery is fleshed out before or during resolution.

Down to 37 regressions.