Update NewString memleaks tests for AMM.
The new algorithm for inserting autodestroys tries to place them as late as possible in
the function, ignoring scoping.  So in testing for memleaks, only a function return
(lexically) will force memory allocated with that function to be released.

So I modified the tests to no longer expect memory to be released at a lexical scope
boundary, by using simple helper functions.

Down to 26 regressions.