Modify deleteUnreachableBlocks() to ignore SymExpr fields in forLoop constructs.
Since there can be initialization code in a for loop that includes declarations, it is
invalid to remove the entire for loop if it is found that the evaluation of the loop index
lies in an unreachable block.  This is true because that evaluation is part of the
condition block, not part of the initialization block.

I am not sure that the same is true for the iterator field.  It seems that when it gives
rise to executable code, this is put in the initialization block preceding the loop.  If
that is the case, then the iterator function call being unreachable means that the entire
loop is unreachable.  However, this decision is complex enough, it should be moved out to
a separate cleanup pass.  The lowest-impact solution to get early unreachable block removal
working is to leave the iterator SymExpr in the tree, just like the index symexpr above.

Without this change, early removal of unreachable blocks causes
users/vass/crash6defPointNotInAST to fail.