Lower stack size for parallel/begin/dinan/mvm_coforall
parallel/begin/dinan/mvm_coforall has been sporadically segfaulting or timing
out for linux32. The test has a big-ish coforall which means a big-ish number
of tasks each with an 8MB stack. Depending on timing it seems the test either
runs out of memory, or uses a lot of swap which drastically increases execution
time.

Lowering the stack size seems to have resolved the regression: tested with 1000
trials.