Fix dygraph zooming
When we zoom, we round the start and end dates so that they always lie exactly
on a date boundary. This makes the graphs cleaner since we only have 1 data
point per day, and it allows us to store the date in the url in a human
readable format.

This rounding was broken for the graph that was actually zoomed on. When you
zoom on a graph, we then apply the same zoom level on other graphs. However the
graph that was originally zoomed in on wasn't having it's start and end date
rounded correctly.