Track processes with highest cpu usage for performance runs
After each trial of a performance run, print out 5 processes with the highest
cpu usage. This is to try and track down performance mysteries that may be the
result of other processes interfering.

Current performance mysteries are --no-local versions of lulesh, for vs while
time comparisons and possibly some tests whose performance differs based on how
long it's been since the machine was rebooted.

This isn't perfect since for long running tests it can miss interference, but
it's a start.