Improve WhileStmt::checkConstLoops().
The check function contained a mixture of structural tests and a semantic test.  As the
routine was written, Both would result in an assertion error.

The structural tests determine whether the check code itself is correctly interpreting the
structure of the tree.  The semantic test determines whether the code represents a loop in
which the condition variable is never updated.  The first set of tests should result in an
internal error.  The second test is appropriately rendered as a user-level warning.

I separated out the two kinds of tests, and added helpful error messages for the caes
where the structural tests are violated.  It should be noted that future changes to the
structure of the AST may lead to violations of some of these tests, in which case the
appropriate course of action is to revise the test to allow for the revised structure of
the tree.  None of the structural tests is set in stone.

Added comments to checkConstLoops().  Updated test_return1.bad so it expects the new warning.