Fixing over-sensitivity to output code format in testmunge.prediff
My previous prediff script for this test was too specific to a given
format of generated code which caused it to break for --no-local and
gasnet testing.  Here, I'm making the test more general.  Previously,
for a given variable 'x' I was checking for 'x ' and 'x_chpl ' in the
generated code depending on whether the --munge-user-idents flag was
set or not.  Now I'm just checking for the presence/absence of
'x_chpl' in the code depending on the state of the flag which should
be much more robust.