Fix to baseline munge user identifier regression
In baseline mode, one of my munge tests failed because the compiler
creates procedures (getters, autodestroys) with the names of the
identifiers and then '_chpl' gets tacked onto the end of them.  It
seemes to me, philosophically, that compiler-introduced identifiers
like this should start with 'chpl_' and be marked with
FLAG_COMPILER_GENERATED, so this fix:

1) Makes the munging logic skip over identifiers with
   FLAG_COMPILER_GENERATED and

2) Adds the flag to getters (while also changing their unique prefix
   from merely '_' to "chpl_"

It is likely the case that there are other compiler-generated symbols
that either don't have the "chpl_" prefix and/or are not marked with
FLAG_COMPILER_GENERATED.  I propose we improve these cases as we come
across them rather than trying to hunt them down here.