Make compiler-generated getters start with 'chpl_'
Prior to this commit, compiler-generated getters started with '_...'
which we're trying to move away from since there's no guarantee that
these identifiers are safe in C.  Changing them here to 'chpl_...'
to match other situations.  This required an update to one .good
file that was sensitive to the names of compiler-introduced getters.

While this change is good housecleaning, it's specifically motivated
by the baseline regression related to munging last night.
Specifically, a subsequent commit will not munge identifiers that
already start with 'chpl_' which will fix a baseline regression.