Add field destructors to freeIterator.
Added freeIterator to the list of functions that are repopulated in the lowerIterators
pass, and added field destructors.  This frees resources that are inserted into an
iterator class when it is created.  Only the fields populated by getIterator are freed.
There are other fields in the iterator class that are used to capture the state at the
iterator call site.  Freeing those will be the subject of a future patch.

Also, for CForLoops, freeHeapAllocatedVars() was inserting frees of HAVs allocated within
the body of the loop in the wrong place.  They need to be freed at the end of the
increment clause.  Otherwise, if they are referenced in the increment clause, there is a
use-after-free error.