Turn off generation of what was previously CHPL_DEVELOPER only output
Previously, we would output much more of the compiler generated structure if the
process running had CHPL_DEVELOPER=true set.  This turned out to be quite a
hassle for us as we were working to expand the chpldoc feature.  For now, make
this output squashed entirely, with a note at each location so that when we
make a flag to enable this output, we know where to add checks for it.