Add code to manage "nude" array and dom objects in getIterator() and freeIterator().
+++ b/compiler/AST/iterator.cpp
Updated rebuildGetIterator to insert an autoCopy call for nude dom or array objects that
are inserted into a field in the iterator class.  Since these types are added to the
autoDestroyMap, the code for freeIterator does not have to be updated, and will handle
this case automatically.

Added a note at the bottom about why this is a special case and how it is handled.

+++ b/compiler/include/type.h
+++ b/compiler/AST/type.cpp
Added new predicates isDomImplType() and isArrayImplType().

+++ b/compiler/resolution/functionResolution.cpp
Updated resolveUninsertedCall() to be immune to the insertion of autoCopy temps, etc.
Added domImplTypes and arrayImplTypes to the list of types for which we resolve the
autoCopy and autoDestroy functions.

+++ b/modules/internal/ChapelArray.chpl
Removed the crufty one-sided refCount increment for nude arrays and domains.  This is now
handled in a more disciplined way in rebuildGetIterator() and (automagically) in
rebuildFreeIterator().

+++ b/modules/internal/ChapelBase.chpl
Added pragam "compiler generated" to chpl__autoDestroy(x: object) so that autoDestroy
functions for instantiated (generic) types will rank above it in resolution.

+++ b/modules/internal/DefaultRectangular.chpl
+++ b/modules/dists/BlockCycDist.chpl
+++ b/modules/dists/BlockDist.chpl
+++ b/modules/dists/CyclicDist.chpl
+++ b/modules/dists/DimensionalDist2D.chpl
+++ b/modules/dists/PrivateDist.chpl
+++ b/modules/dists/ReplicatedDist.chpl
+++ b/modules/internal/ChapelArray.chpl
+++ b/modules/internal/DefaultAssociative.chpl
+++ b/modules/internal/DefaultOpaque.chpl
+++ b/modules/internal/DefaultRectangular.chpl
+++ b/modules/internal/DefaultSparse.chpl
+++ b/modules/layouts/LayoutCSR.chpl
Added explicit autoCopy and autoDestroy functions for "nude" domain and array types.