Comment out the change and restore the previous status.
Regular testing of the auxiliary file systems is farther in the future than was
originally thought when this change was made.  This change is still deemed
important, but so is maintaining our current working implementation.  Since
we don't have a set up for testing to ensure this change does not break the
implementation before committing, merging the change is not reasonable.  Thus,
I have chosen to comment out the change, with a note so that someone working on
the auxiliary file system implementation in the future will know to reimplement
and test the change, hopefully leading to its eventual merger into master.  I
left a date by the notes for posterity.