Update use of macros for atomic uintptr_t functions
For some reason that baffles me using the DECLARE_ATOMICS marco for uintptr_t
causes build issues on netbsd, but calling the 3 macros that it does works
fine.

This seems like an ok, though unfortunate workaround. So far as I can tell we
don't actually use any of the atomic operations on uintptr_t either in the
modules or runtime so it might be easy enough to just remove the declaration of
it.

I wanted to wait and talk to Michael about that, but this seems like an ok fix
in the meantime.