Now it works!
Part of the change I made assumed documentation for type fields would be stored
in the TypeSymbol for the type.  This was unnecessary, *and* as a bonus it
introduced bugs.  Glad I spent all day on that.

Also added a future for usesTypeVar, as it generates .(this, "t") as output
instead of just t.