Update functions to get memory and cpu/pu count in chplsys for netbsd
We already had implementations for getting the amount of memory on a machine.
Just needed to include sys/sysctl.h to get those routines to work again. We
didn't have implementations to get the number of physical and logical cores.
However the logic we use for linux (getting it from /proc/cpuinfo) works on
netbsd.

Getting the available cores/pus based on accessibility information
(sched_getaffinity) is different than linux and I didn't implement that for
netbsd.

It's similar-ish to how it's done for linux:

Instead of cpu_set_t there is a cpuset_t which has to be created with
cpuset_create()/cpuset_destroy(). The call is sched_getaffinity_np() instead of
sched_getaffinity() and CPU_COUNT isn't implemented. I based my initial attempt
off of code in hwloc-1.9.1/src/topology-netbsd.c, but wasn't sure how to get
around CPU_COUNT not working so I gave up for now. I can work on this later,
though I am starting to feel we should rely on using hwloc for these types of
questions.