Capturing a bug that Thomas identified as a future
While working on chpldocs, Thomas noted that compiling the NewString
module causes a segfault.  The reason for this is subtle and unlikely
to be something that an end-user runs into.  It essentially relates to:

(a) naming on the command line a file that defines a module with the
    same name as an internal/standard module and

(b) the internal/standard module being defined in a file that defines
    multiple modules.

The fix is probably not too difficult, but I didn't undertake it today
in order to focus on higher priority work.  In the specific instance
of chpldoc and NewString, I'm betting that NewString will go away as
new strings come online and/or not be something we want to document
before they do (at which point, we may want to update this test to
point to another such internal/standard module if one can be found).

I've put some notes about how to fix this bug in the .future file.