When "erasing" comment prefixes, do not erase more chars than are in string.
For example, if line is empty, trying to "erase" three characters will result
in an exception in some compilers.

It appears that different STL implementations treat erasing non-existent
characters in different ways. Some consider it an error and raise, while other
silently ignore it. This solution should work regardless of STL implementation.