Add tests to ensure a method defined outside of a class is displayed outside
This test establishes that methods on a class or record that are defined outside
of the initial class definition should be output at the module level rather than
at the class/record level, but should still indicate that they attached to the
data structure in question.

Current incorrect behavior: these methods are printed twice, once at the class
or record level, and once at the module level, but the module level output does
not indicate its connection to the class.