Minor fixes to documentation; remove Random from spec
In "documenting" I renamed some things that broke some things in
obvious ways.  Fixing those here.

I also took this opportunity to pull the Random module out of the spec
(finding, in doing so, that it didn't match the implementation
anymore... go figure).  And, also found a few notes that didn't occur
to me in writing up new documentation, so updated the comments to
reflect that.