Move deallocation of heap-allocated variables to increment clause of a CForLoop.
In a C-style for loop, the code in the increment clause can depend on variables that are
declared within the body of the loop.  To avoid read-after-free errors, the freeing of
heap-allocated variable is moved to the end of the increment clause.

This sometimes leads to rather "fat" for-statements in the generated C output, but our
code generator does the right thing and it is still legal C code.  We could make the code
a bit prettier by enumerating the heap-allocated variables that are actually accessed in
the incr clause and then moving the free() calls for the rest of them back to the end of
the loop body.

[x] Full regression test run.