Fix the failing standard test case for FLAG_METHOD applied to secondary methods
It was discovered that the flag FLAG_METHOD was only applied to methods defined
within a class definition, not those defined outside of the class definition.
This struck us as an oversight, so we added a separate flag,
FLAG_METHOD_PRIMARY, to handle the cases where we wanted to distinguish between
these cases, and replaced those checks with checks to this new flag.  In the
case of isClassMethodCall(), used by insertNilChecks() to determine if we should
check the type instance used for the call is not nil, it seemed that what was
really intended was to check against all methods rather than merely primary
ones.  However, allowing this check to expand encountered a bug with --no-inline
and --no-local: widened strings were now matching though strings are not
classes.  Since it doesn't make sense to check against widened strings, I added
an additional check to ignore strings in isClassMethodCall.