Use multi-configurations for compiler performance graphs
Beef up multi-configuration support (e.g. local vs no-local) and make compiler
performance graphs use multi-configurations.

The point of this was to beef up multi-configuration support so we can start
using it for things like:
 - local vs. --no-local
 - gnu vs. cray vs. intel vs. pgi
 - flat vs. numa
 - ugni vs. gn-aries vs. gn-mpi

Multi-configuration was only used previously to compare fifo vs qthreads and it
was a hack thrown together that was pretty buggy and slow. I switched compiler
performance graphs over to use multi-configurations because it was a good test
case to make sure multi-configurations is working well.

With this, multi-configurations work well for up to 2 configurations. A small
upcoming patch will extend this to an arbitrary number (the current sticking
point is we only have 2 line types (normal and dashed) for dygraphs and it's
hard to distinguish the other configurations without different line types.

Beyond beefing up multi-configuration support, this cleans up compiler
performance graphs pretty nicely, reduces their memory footprint online from
~100MB to ~35MB and makes the graphs faster.

More details:

We graph the compilation times for 'all' tests and for 'release/examples'
tests. This is meant so we can see the average pass times in general, but also
see the average times for 'real' world tests like most of the tests in
release/examples.

Previously we did this by duplicating the keys and files in the .graph files
and special casing code in our javascript that creates the dygraphs. This moves
them to use the multi-configuration functionality and beefs up that
multi-configuration support in order to do so.

File specific changes:

test/performance/compiler/*.graph:
  - Remove duplicate keys for '(all)' and '(examples)', handled with multi-conf

util/start_test:
  - Calls gengraphs with '-m all:v,examples:v' to use multi-conf:

util/test/combineCompPerfData:
  - Modify where we place the average timings for passes. multi-conf expects
    each configuration to live in a sub-folder that has the same name as the
    configuration.

util/test/perf/perfgraph.js
 - Allow multi-conf state to be stored and retrieved from the URL and add
   helper functions to do that
 - change use of expand to be a boolean now

util/test/genGraphs:
 - Rename multiDesc to multiConf since that's a better name.
 - Handles stripping series in gengraphs instead of the javascript (used for
   showing top X compiler passes.) This is needed because previously we would
   just hide the other series in dygraphs, but that conflicts with how we
   toggle different series.
 - Change expand to a boolean. Previously it took the number of series to
   expand, but numseries specifies how many series to strip in gengraphs and
   expansion just expands all series (which means it can be simplified in the
   future.)
 - Beef up multi-configuration support:
    -  Update how configurations are specified to gengraphs. It used to be that
       each configuration had to be set as a different arg and the one
       specified first was the only one displayed by default. Now the
       configurations are displayed as a comma separated list. each
       configuration can have a ':v' after it to denote it should be visible by
       default. This new form also makes adding other options in the future
       easier. For instance it makes it easier to add the ability for the
       configuration and the sub-folder to find the .dat files in be different.
    - Improve how we find the .dat files each configuration. Previously if a
      .dat file for a configuration couldn't be found we ignored it and
      dygraphs blew up. This updates that so if a .dat file wasn't found we
      assume the test wasn't run for this configuration or timed out. We still
      need to generate a series for dygraphs so we try to find the .dat file
      from one of the other configurations and copy the series. This way
      dygraphs is happy even if there is no data for a configuration. This is
      needed for things like ganset-aries vs gasnet-mpi since in many cases
      gasnet-mpi won't have any data since it timed out.