Reinstitute is*Method() methods on FnSymbols, and use them.
As Mike pointed out, though I was the only one using them, it's better style to
define an abstraction that we can change at whim and which clearly delineate
the hierarchy between methods, primary methods, and secondary methods.  So the
previous commit reverts their removal and this one updates specific places with
their usage.