Allow multi-configs perf graphs to work with any number of configs
Previously multi-configs graphs only really worked with 2 configurations
because it relied on hacky behavior that was used for compiler performance
graphs. There was a special setColors function that made the strokePattern and
color the same for all configurations of a series. This function was expensive
and was klunky since it was originally used when we sorted the graphs online
instead of in gengraphs. This setColors also only worked for two configurations
and had to be called after the graph was generated causing an expensive
re-render

This commit adds two functions which generate a list of colors and stroke
patterns to use for a graph and they work for an arbitrary number of
configurations. They only need to know the series for the graph, so they can be
called before the graph is created, which prevents the expensive re-renders.

Getting rid of setColors() allowed me to remove the hack to set primaryString
and secondaryString which also allowed graphExpansion to be simplified.