Treat the RVV as always belonging to a separate clique.
In extractSymbols, don't let the return value variable join any alias cliques.  In
processBitwiseCopy, reset ownership bits if ownership passes from one clique to another.
(This should only happen for the RVV at present, but is written to work generally.)

Revised isBitwiseCopy to return more information: which argument the given SymExpr is in a
bitwise copy.

Down to 28 regressions.