Futurize some tests that require out-of-scope fixes.
Correct results from the opEquals test depends on the operation of
changeRetToArgAndClone() pulling the return value autoCopy into the called function.  This
is behavior we don't want to rely on.  The problem is that the initializer built by
chpl__buildAssociativeArrayExpr and the array into which it is copied both count as arrays
in the underlying dom.  The trunk version succeeds because the autoDestroy of the
initializer is called inside of the build routine, whereas in the AMM version both the
unnamed initializer array and the named array into which the initializer is copied are
both live until the end of the main routine.

The other tree tests are failing because the parloopexpr used to initialize an array
within a record refers to the record as a whole -- which breaks one of the rules in the
new constructor story.  (All fields must be initialized before any method can be called on
the record as a whole.)  New AMM rules require that the record is copied when it is assigned
to a field in an iterator.  When this happens in the iterator defining the function of the
parloopexpr, it leads to infinite recursion.  Creation of the record copy in the parloop
iterator can also reference fields in the partially-initialized record that have not been
initialized yet.  In two different ways, these tests are expected to fail.

Details are provided in the added .future files along with recommendations on how these
implementation problems can be addressed.