Fix the zero array reference errors in the CSR layout tests.
I missed adding constructors and destructors for the CSRArr, so I added those.  Two more
regressions resolved.

I also added code for automatically adding an autoCopy call in rebuildIterator(), but this
turned out to be the wrong solution.  Nonetheless, some of the added code provides a
slight factoring.  I left the rest of it in a #if 0 block, in case it is needed later.  It
can (and should) be cleaned up before the commit to trunk.