Enable one-sided MM for iterator records (workaround)
To resolve regressions in studies/hpcc/common/bradc/unit/, I added code to bump the
reference count on arrays and doms that are copied into an iterator record.  For the test
cases that fail -- leadFollowOpt and parBlock1D -- this causes the iterator containing a
BlockDom returned by BlockDom.these() to have its refcount increased from 0 to 1 before
further operations are performed upon it.

Without this, the corresponding getIterator() bumps the refcount and the corresponding
freeIterator() decrements the refcount and deletes the dom as the refcount transitions to
zero.  Further references to the dom contained in the IR will reference data that has
already been freed.

In summary, when an iterator record is created from a nude array or dom that initially has
a refcount of zero, it needs to increment the refcount in those objects.  An alternative
would be to have the "these" function that returns the nude array or dom increment the ref
count.  That is actually where we started.  It fails because nude doms and arrays are
class instances, so there is no (standard) way to capture references to them and ensure
that they are later destroyed.

Typically, when handling class instances, it is up to the client code to perform memory
management.  So the approach of having the iterator record manage references to nude doms
and arrays is consistent with the general programming model.