Update some .bads and fix some .goods.  Also, remove a .future file
The generic-parenthesesless-*.goods were expecting an empty class to be output,
but since they were calling the parentheses-less method, what should be expected
is the output of the method itself.  After my change to methods defined outside
of the class declaration, these should pass, but I'm going to let them run
tonight just to be sure.

generic-parenthesesless-big1.future started passing as a result of my change,
so remove the .future file.

functions/kbrady/proc_scoping and modules/bradc/useFileSystemShadowsPath started
failing in a different way, but the new output seems reasonable, so update the
.bad to track it.