Fix breaking distributions test, test_block_representation.
With gasnet, this test was failing because a nil check was inserted before the
writeThis method on _ddata was called.  We agree that methods should not be
called on nil instances, and we decided that if the _ddata instance is nil it
should be output as nil.  To that purpose, I have expanded the nil output case
to include _ddata and updated the expected output accordingly.