Two new tests of passing arrays to extern C routines expecting arrays
multidimArrToExtern.chpl is a test that passes local multidimensional
arrays to an external C routine; prior today's work, this was not
supported, but not for any good reason that I can tell.

sparseToExtern.chpl passes a sparse to an extern (not currently
supported), and is designed to demonstrate better error messages that
I implemented today by using compilerError()'s optional depth
argument.