Extend c_ptrToArr() to support multidimensional arrays
It's never been clear to me why c_ptrToArr() was restricted to only
apply to 1D arrays given that it seems to work for multidimensional
as well.  Quick tests indicate that this is the case.

While here, noted that the error messages for arrays that are not
supported can be significantly improved by pointing up a level of
depth in the compilerError() routine.

In a separate commit, added tests that exercise both of these changes.