Decode our perf graph query string before looking at it
Our query string uses ','s and '/'s to store lists of graphs/configurations and
store the startdate and enddate so that we can share graphs with each other.
One could argue that we should % encode them since they are technically still
reserved characters in the query string even though they have no reserved
purpose for us. % encoding is used because certain characters have reserved
purposes in the URI so you need some way to encode them in URI without actually
using the characters. e.g. '/' is encoded as '%2F'.

From a percent-encoding article: "In the 'query' component of a URI (the part
after a ? character), for example, '/' is still considered a reserved character
but it normally has no reserved purpose, unless a particular URI scheme says
otherwise. The character does not need to be percent-encoded when it has no
reserved purpose."

Some other sites where we might post perf graph links (i.e. facebook) have
reserved purposes for some of the characters in the query string, so they
percent encode our query string.

Previously we didn't decode the string so our graphs couldn't parse the
encoded date. This just updates our graphs to use jquery's URI decoder before
we look at the query string. We don't have to encode the query string ourselves
since we have no reserved uses of the reserved symbols and I *think* any site
we post to will do the encoding themselves if they use any of the reserved
characters. We could always encode it just to be safe, but that makes the URI
uglier