Add capability to only test futures that have skipifs
For a while we've wanted a testing mode where futures are run if and only if
there's a skipif for the test. Historically there was no middle ground and we
either ran 0 or all ~800 futures. We turned futures off for most configurations
since we didn't want to waste time running them everywhere with little reward.
However, some futures are platform specific and we would never have noticed if
they started passing.

This allows us to test configuration specific futures, without have to run all
the futures. There are only ~60 future tests that also have .skipifs. I would
guess that maybe 5-30 futures will run for any given configuration so it
shouldn't add much overhead to testing.

The current default for nightly is to only test futures with skipifs. You can
toss -futures to test all futures, or -no-futures to skip all futures.

Only the "standard" linux64 configuration is tossing -futures for now. I'll
ping the group to see what other configurations we might want to add. I thought
about gasnet-everything and darwin, but wasn't sure.

I was originally thinking that -no-futures wouldn't be used much if at all now.
I removed all uses of it, but now I'm wondering if it makes sense to use it for
verify and memleaks still.