Allow the _real binary to be wrapped for most launchers
Add an undocumented function chpl_get_real_binary_wrapper() which returns the
value of CHPL_LAUNCHER_REAL_WRAPPER if set, otherwise it returns "". This is
being added so we can use our whole program execution timers for xc performance
testing. It's not explicitly called CHPL_LAUNCHER_TEST_EXEC_TIMER or something
because it could be used more generally. Still, it doesn't work for all
launchers so I don't really want to advertise it (at least not yet.)

Without this we're including the time to get the resources we requested and any
pre-exec node setup for xc performance testing (and for any performance testing
that might use a queued launcher.) This makes whole program execution widely
inaccurate especially since we have multiple jobs running concurrently that
request the same nodes.

This will work with slurm, pbs-aprun, and aprun. It will "work" for any other
launcher that uses chpl_bundle_exec_args(), but I don't think that includes
many other launchers.

This doesn't really work with the amudprun launcher since the gasnet launcher
adds args to first non-positional argument that gasnet then uses for setting
up. We could make the timers take an optional argument that specifies which
argument is the binary, and reposition any other args to go after the binary.
There's no real need for that yet, since we don't do performance testing with
amudprun and even if we did we're not waiting on a reservation, so timing the
non _real won't be far off.