Time the _real binary for perf testing when using a supported launcher
For xc performance testing, our whole program execution timings are way off
since they're timing the executable that launches the job, not the _real
executable. This updates sub_test to time the _real binary using the new
CHPL_LAUNCHER_REAL_WRAPPER env var for launchers that are known to support it.
Currently that's pbs-aprun, aprun, and slurm-srun which is all the ones we
might care about in the near future.

Some details:
 - Updates how we set the timer if one of the above launchers is used
 - move launchcmd setup out of conditional nest so that whole program execution
   can be used at the same time as launchmd.
 - Swap the order that we add the launcher timeout. For whole program execution
   not using launchcmd, cmd will be the timer, and args will be the binary. The
   timeout needs to go after the binary. I think this means that launchcmd
   doesn't really need to support --walltime anymore since --walltime is always
   given to the executable now, but it doesn't hurt to leave it in.