Add debugging printfs for incrementing and decrementing distribution reference counts.
+++ b/modules/internal/AtomicsCommon.chpl
Added a wrapper function for the atomic read().

+++ b/modules/internal/ChapelBase.chpl
Added the config param debugDistRefCount.
It has to go in ChapelBase as a workaround for the false-positive use-before-def-error.

+++ b/modules/internal/ChapelDistribution.chpl
Narrowed the range of valid distribution reference counts to [0,1000].  The upper end is
arbitrary, so should be made a config param.  A number comparable to a small multiple of
the number of locales is reasonable; (uint64_t)(-1) most definitely is not.  (For that
matter, any stack or heap address converted to an integer is likely to be bogus.)